### 2025-04-20 #7967 @Soontosh
The library is based on a research project we've been working on. We developed a simplified and more general-purpose version of the architecture we've developed for use by other researchers and devs. We've taken our jupyter notebooks for the project and converted it into a nice easy to use Python library. Happy to provide more information if needed.
