### 2025-04-29 #7967 @Manvi-Agrawal
Hi @crvernon, thanks for providing me this opportunity. I am currently busy and would be able to review it a week later or so, if that's okay.

Meanwhile it would be helpful if @builtbypyro  could you go through the JOSS reviewer checklist and see that all of the items mentioned there are present in this submission.  From a brief look I couldn't find the link to software's API documentation website which is required for joss submission.
