### 2025-05-30 #8320 @editorialbot
[REVIEW]: LLMOrchestrator: A Multi-Model LLM Orchestration Framework for Reducing Bias and Iterative Reasoning
**Submitting author:** @builtbypyro (<a href="http://orcid.org/0009-0009-7588-7498">Trisanth Srinivasan</a>)
**Repository:** https://github.com/builtbypyro/LLMOrchestrator.git
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
