### 2025-06-12 #8320 @sumansuhag
LLMOrchestrator tackles the need for multiple agents to work together with language models while keeping trust in mind. It has solid benchmarks, good visual aids, and info on how to scale it, making it a useful resource for anyone concerned about AI safety and quality.

Want me to put together a comment or summary to share on GitHub or JOSS?
