### 2025-06-19 builtbypyro/LLMOrchestrator#8 @Manvi-Agrawal
ERROR running LocalModel example
### Context
As part of JOSS review, Tracking in https://github.com/openjournals/joss-reviews/issues/8320

Machine: Windows OS 11
TODO: Add Python version details

### Code

```python
from LLMOrchestrator.models import OpenAIModel, LocalModel
from LLMOrchestrator.controller import Controller, PromptTemplate

# Initialize models
generator = LocalModel(
    model_name="facebook/opt-125m",
    device="cpu"
)

verifier = LocalModel(
    model_name="facebook/opt-125m",
    device="cpu"
)

# Create a prompt template
template = PromptTemplate(
    "Please provide a detailed analysis of: {prompt}"
)

# Initialize controller
controller = Controller(
    generator=generator,
    verifier=verifier,
    max_iterations=3,
    max_verifications=2,
    parallel_processing=True,
    cache_enabled=True,
    adaptive_learning=True,
    monitoring_enabled=True,
    prompt_template=template
)

# Execute with a prompt
result = controller.execute(
    prompt="Analyze the impact of artificial intelligence on healthcare.",
    stop_early=False
)

# Get performance metrics
metrics = controller.get_validation_metrics()
print(f"Quality Score: {metrics.quality_score}")
print(f"Confidence: {metrics.confidence_score}")
print(f"Processing Time: {metrics.processing_time}s")
```

### Stacktrace
2025-06-19 13:06:02,341 - INFO - Starting execution for prompt: Analyze the impact of artificial intelligence on h...
2025-06-19 13:06:02,341 - INFO - Starting iteration 1/3

ITERATION 1/3
   Generating response... (timeout=30s)
Both `max_new_tokens` (=50) and `max_length`(=100) seem to have been set. `max_new_tokens` will take precedence. Please refer to the documentation for more information. (https://huggingface.co/docs/transformers/main/en/main_classes/text_generation)
   Generated: The world is changing. But what is the impact?

I have been following the healthcare sector for a nu
   Verification attempt 1/2
2025-06-19 13:06:09,290 - ERROR - Verification error: cannot unpack non-iterable float object
   Verification attempt 2/2
2025-06-19 13:06:09,353 - ERROR - Verification error: cannot unpack non-iterable float object
2025-06-19 13:06:09,353 - WARNING - No valid refinement achieved in this iteration
   No valid refinement achieved in this iteration
2025-06-19 13:06:09,353 - INFO - Starting iteration 2/3

ITERATION 2/3
   Generating response... (timeout=30s)
Both `max_new_tokens` (=50) and `max_length`(=100) seem to have been set. `max_new_tokens` will take precedence. Please refer to the documentation for more information. (https://huggingface.co/docs/transformers/main/en/main_classes/text_generation)
   Generated: The healthcare industry is in crisis. With the recent emergence of artificial intelligence (AI) and
   Verification attempt 1/2
2025-06-19 13:06:14,092 - ERROR - Verification error: cannot unpack non-iterable float object
   Verification attempt 2/2
2025-06-19 13:06:14,191 - ERROR - Verification error: cannot unpack non-iterable float object
2025-06-19 13:06:14,191 - WARNING - No valid refinement achieved in this iteration
   No valid refinement achieved in this iteration
2025-06-19 13:06:14,191 - INFO - Starting iteration 3/3

ITERATION 3/3
   Generating response... (timeout=30s)
Both `max_new_tokens` (=50) and `max_length`(=100) seem to have been set. `max_new_tokens` will take precedence. Please refer to the documentation for more information. (https://huggingface.co/docs/transformers/main/en/main_classes/text_generation)
   Generated: In a recent article in the Journal of Medicine, researchers at the University of California at San F
   Verification attempt 1/2
2025-06-19 13:06:19,268 - ERROR - Verification error: cannot unpack non-iterable float object
   Verification attempt 2/2
2025-06-19 13:06:19,363 - ERROR - Verification error: cannot unpack non-iterable float object
2025-06-19 13:06:19,363 - WARNING - No valid refinement achieved in this iteration
   No valid refinement achieved in this iteration
2025-06-19 13:06:19,363 - WARNING - No valid outputs generated after all iterations
Traceback (most recent call last):
  File "LLMOrchestrator\examples\quickstart.py", line 39, in <module>
    result = controller.execute(
             ^^^^^^^^^^^^^^^^^^^
  File "LLMOrchestrator\controller.py", line 354, in execute
    raise RuntimeError("No valid outputs generated after all iterations")
RuntimeError: No valid outputs generated after all iterations
