### 2025-07-08 adf44d5
Bump transformers from 4.50.3 to 4.51.0
Bumps [transformers](https://github.com/huggingface/transformers) from 4.50.3 to 4.51.0.
- [Release notes](https://github.com/huggingface/transformers/releases)
- [Commits](https://github.com/huggingface/transformers/compare/v4.50.3...v4.51.0)

---
updated-dependencies:
- dependency-name: transformers
  dependency-version: 4.51.0
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
