### 2025-07-08 3c636c7
Bump transformers from 4.50.3 to 4.52.1
Bumps [transformers](https://github.com/huggingface/transformers) from 4.50.3 to 4.52.1.
- [Release notes](https://github.com/huggingface/transformers/releases)
- [Commits](https://github.com/huggingface/transformers/compare/v4.50.3...v4.52.1)

---
updated-dependencies:
- dependency-name: transformers
  dependency-version: 4.52.1
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
