+++ requirements.txt
@@ -30,7 +30,7 @@ sympy==1.13.1
 tokenizers==0.21.1
 torch==2.6.0
 tqdm==4.67.1
-transformers==4.50.3
+transformers==4.52.1
 typing-inspection==0.4.0
 typing_extensions==4.13.0
 urllib3==2.3.0
