### 2025-08-05 #8320 @pebabion
@crvernon I provided the review earlier with with OpenAI API. I can review the Local Model and post it here within this week.
