+++ LLMOrchestrator/controller.py
@@ -93,24 +93,33 @@ class PromptTemplate:
 
 class OutputCache:
     """Enhanced caching system with quality-based retention."""
-    def __init__(self, cache_dir: str = ".cache", max_size_mb: int = 1000):
+    def __init__(self, cache_dir: str = ".cache", max_size_mb: int = 1000, auto_save: bool = False):
         self.cache_dir = Path(cache_dir)
         self.cache_dir.mkdir(exist_ok=True)
         self.cache_file = self.cache_dir / "output_cache.json"
         self.max_size_bytes = max_size_mb * 1024 * 1024
         self.cache: Dict[str, Dict[str, Any]] = self._load_cache()
         self.lock = threading.Lock()
+        self.auto_save = auto_save  # Only save automatically if enabled
+        self.dirty = False  # Track if cache has unsaved changes
     
     def _load_cache(self) -> Dict[str, Dict[str, Any]]:
         if self.cache_file.exists():
-            with open(self.cache_file, 'r') as f:
-                return json.load(f)
+            try:
+                with open(self.cache_file, 'r') as f:
+                    return json.load(f)
+            except:
+                return {}
         return {}
     
     def _save_cache(self):
         with self.lock:
-            with open(self.cache_file, 'w') as f:
-                json.dump(self.cache, f)
+            try:
+                with open(self.cache_file, 'w') as f:
+                    json.dump(self.cache, f)
+                self.dirty = False
+            except Exception:
+                pass  # Silently fail on cache save errors
     
     def get(self, key: str) -> Optional[Dict[str, Any]]:
         with self.lock:
@@ -121,7 +130,9 @@ class OutputCache:
             if self._get_cache_size() > self.max_size_bytes:
                 self._prune_cache()
             self.cache[key] = value
-            self._save_cache()
+            self.dirty = True
+            if self.auto_save:
+                self._save_cache()
     
     def _get_cache_size(self) -> int:
         return sum(len(str(v).encode('utf-8')) for v in self.cache.values())
@@ -224,3 +235,213 @@ class Controller:
             return
         self.cache.set(prompt, {
             'output': output,
+            'timestamp': time.time(),
+            'metrics': self.metrics.__dict__ if self.metrics else {}
+        })
+
+    def _generate_cache_key(self, prompt: str) -> str:
+        """Generate a cache key for the given prompt."""
+        import hashlib
+        return hashlib.md5(prompt.encode()).hexdigest()
+
+    def execute(self, prompt: str, **kwargs) -> str:
+        """
+        Execute the orchestration process for a given prompt.
+        
+        Args:
+            prompt: The input prompt to process
+            **kwargs: Additional execution parameters
+            
+        Returns:
+            str: The generated and verified output
+        """
+        start_time = time.time()
+        
+        # Check cache first
+        cached_output = self._get_cached_output(prompt)
+        if cached_output:
+            if self.logger:
+                self.logger.info(f"Retrieved cached output for prompt: {prompt[:50]}...")
+            return cached_output
+        
+        # Apply prompt template if available
+        if self.prompt_template:
+            prompt = self.prompt_template.get_template().format(prompt=prompt)
+        
+        # Determine iterations
+        iterations = self.max_iterations
+        if self.dynamic_iterations:
+            iterations = self.dynamic_iterations(prompt)
+        
+        # Generate and verify output
+        output = None
+        refinement_count = 0
+        
+        for i in range(iterations):
+            try:
+                if self.logger:
+                    self.logger.info(f"Iteration {i+1}/{iterations}")
+                
+                # Generate output
+                output = self._process_with_retry(
+                    self.generator.generate_output,
+                    prompt
+                )
+                
+                # Verify output
[…]
