### 2025-10-14 4e6bc8d
Fix bug with the usage examples
