### 2025-10-16 #8320 @sumansuhag
My understanding is limited to what I’ve been experimenting with Cursor
IDE’s coding agent integrated via MCP to SGraph models. I think this
approach can be generalized.

Here is what I’ve done.

   1. Develop MCP server for utilizing local large graphs.
   2. Setup an analysis process that provides continuously new versions of
   the graph (if there are changes in the source). The most recent model is
   made conveniently available via symbolic link.
   3. The agent has been configured to use SSE, but also Stdio approach
   would work in this case. The sgraph-mcp-server has been extended with
   suitable tools to make AI agent’s “life” more easy.
   4. When the end user is discussing with the agent, the agent knows the
   MCP tools that are available, and it understands quite well when the end
   user suggests utilization of graph RAG, i.e. making tool calls to the
   graph. Currently there aren’t yet any notification about available new
   graph in the sgraph-mcp-server but that shouldn’t be a big task to
   implement.
   5. One important aspect is the performance. The graphs may contain
   millions of elements, which makes the loading slow. That’s why the
   sgraph-mcp-server keeps it in memory, and is able to help with traversal
   questions very fast.

I don’t go now for the concrete use cases, as this is very general topic,
but you are welcome to connect with me for more information. The graph
contains also metrics (attributes) which makes the utilization of the
information very interesting. Any attributes can be provided from the
analysis phase, for both nodes and edges.

The real challenge right now is to make the agent understand itself when it
is wise to retrieve information from the graph, if the information can also
be retrieved from other sources (e.g. in coding, it can execute vast amount
of grep commands, waste tokens, use LSPs (usefulness depends on the
programming language) etc.). But obviously in large projects the
information (developers don’t like to keep 10 MLOC code in sync on their
PCs) is not available so the graph RAG is the only way to get sufficient
context information.

This approach is generic and works outside software devevelopment genre,
whenever there are high complexity that cannot be represented as lists.
Basically, ETL process is needed, and the LLM agent needs to be prompted
well, and provided with MCP tooling to be able to traverse the elements.

On Wed, Oct 15, 2025 at 9:27 PM Suman Suhag ***@***.***>
wrote:

> Hey! Bro,
>
> LLMs:
>
>    1. Self-Attention Mechanism: This mechanism enables the model to weigh
>    the importance of different words in a sentence, allowing it to capture
>    relationships between words regardless of their positions in the sequence.
>    2. Embeddings: LLMs represent words as dense, high-dimensional vectors
>    (embeddings) that encode semantic information. These embeddings are learned
>    during training and enable the model to understand the contextual meaning
>    of words.
>    3. Fine-Tuning: After pre-training on a large corpus of text data,
>    LLMs undergo fine-tuning on specific tasks, such as language translation,
>    text summarization, or sentiment analysis. This process adapts the model to
>    perform well on the target task by adjusting its parameters.
>
> Limitations of Large Language Models
>
>    - Most LLM models are trained on data up to a certain point in time,
>    after which the models don’t really have information on what’s happening in
>    the world, they also do not have information about your specific internal.
>    As an example, most of the GPT models are trained on data until Sep 2021,
>    which means if you ask a question about a news event which happened in
>    2023, the model by itself won’t know the factual answer to it, and will
>    likely predict a sequence of words which most closely fit the user input
>    (aka hallucination). The models also don’t have information about your
>    private customer data, internal knowledge bases, proprietary code and can’t
>    answer specific questions about those out-of-the-box. If your use case
>    requires leveraging this data, you’ll need to leverage other mechanisms to
>    provide this information.
>    - LLMs by themselves cannot take actions: It’s important to understand
>    that LLMs are large language models which can take inputs or instructions
>    and return text to satisfy the requested criteria. It is no Skynet. In
>    order to create a Skynet, the LLMs will need a way to reason or think,
>    execute a plan, connect to internal/external systems or services, execute
>    sequential actions. Imagine you want to create an evil Skynet which can
>    launch a rocket by itself. There are several things it will need to do,
>    first, it will need to think and plan which rockets to launch, how to get
>    access, which programs to execute, trigger the programs, if it has to hack
>    the passwords it will have to think of it as a specific action and execute.
>    Good news is that it is indeed possible to build your own Skynet powered by
>    LLMs, keep reading if you want to do that.
>
>
> On Wed, Oct 15, 2025 at 3:05 PM Chris Vernon ***@***.***>
> wrote:
>
>> *crvernon* left a comment (openjournals/joss-reviews#8320)
>> <https://github.com/openjournals/joss-reviews/issues/8320#issuecomment-3408480813>
>>
>> OK, those comments have been removed
>>
>> —
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/8320#issuecomment-3408480813>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/BJBHOMJEU77OQLU743IK6ZL3X3AKVAVCNFSM6AAAAAB6IFJRKWVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZTIMBYGQ4DAOBRGM>
>> .
>> You are receiving this because you commented.Message ID:
>> ***@***.***>
>>
>
