### 2026-01-23 #8320 @crvernon
@Soontosh - per my comment [here](https://github.com/openjournals/joss-reviews/issues/8320#issuecomment-3356854672) about the submitting author being absent from this review:

I don't see where the submitting author has made a single comment in this review thread, but you as the second author have done all of the interaction.  I now also see that the submitting author's GitHub user name `@builtbypyro` no longer exists as a GitHub account.

❓ Is the submitting author still involved and functioning as the lead author; if so, then I need a valid GitHub user name to move forward with and I need the author to make regular updates and responses to inquiries here in this thread.  It is perfectly fine for you to also do so and I appreciate how responsive you have been throughout this review.
