### 2026-03-30 #8320 @pebabion
Hi @crvernon - I believe the latest review from me is still valid. The LocalModel bug has been fixed but the word count (paper length) is still not satisfactory.

Thank you!
