### 2026-03-30 #8320 @Manvi-Agrawal
Agree with @pebabion . Lets fix the word count, and we could look into enhancing the API documentation later, because current api.md covers the major API.
